[chore] 유효성 검사 로직 파일 이동:truck:
