+++ client/register.html
@@ -7,8 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>회원가입</title>
     <link rel="stylesheet" href="./css/pages/register.css" />
-    <script src="./js/common/login.js" type="module"></script>
-    <!-- <script src="./js/common/common.js" defer></script> -->
+    <!-- <script src="./js/common/login.js" type="module"></script> -->
     <script src="./js/common/resgister.js" type="module"></script>
   </head>
   <body>
